+++ .github/workflows/main.yml
@@ -15,7 +15,7 @@ env:
   ENV_NAME: 'dipu_poc'
   SLURM_PAR_CAMB: ${{ secrets.SLURM_PAR_CAMB }}
   CLUSTER_CAMB: CAMB
-  PYTORCH_COMMIT: "0388400f3f8a8ecae2f809ba40ca3ddd5a8b9028"
+  PYTORCH_COMMIT: "680efdb8400bbbf741d6a2e8afaa1fa0bf99f352"
   DIOPI_COMMIT: "7ec02a31440857fed9405fda8d519d4dee8c17e4"
 
 jobs:
@@ -42,23 +42,23 @@ jobs:
           rsync -a --delete /home/autolink/rsync/sourcecode/pytorch ${CLUSTER_CAMB}:${NFS_PATH}/${GITHUB_RUN_NUMBER}/
           rsync -a --delete ${GITHUB_WORKSPACE}/ ${CLUSTER_CAMB}:${NFS_PATH}/${GITHUB_RUN_NUMBER}/source/
 
-  Lint-C-Python:
-    name: Lint c python
-    runs-on: github-poc-ci
-    needs: [Rsync]
-    env:
-      MLU_REQUESTS: 0
-    steps:
-      - name: lint-c-python
-        run: |
-          ssh ${CLUSTER_CAMB} """
-          set -e
-          cd ${NFS_PATH}/ && ls -al && find ${NFS_PATH}/ -maxdepth 1 -ctime 1 -type d |xargs rm -rf
-          source /mnt/lustre/share/platform/env/${ENV_NAME}
-          cd ${NFS_PATH}/${GITHUB_RUN_NUMBER} && rm -rf ${GITHUB_JOB} && cp -R source ${GITHUB_JOB} && cd ${GITHUB_JOB}
-          srun --job-name=${GITHUB_JOB} --partition=${SLURM_PAR_CAMB} -x HOST-10-142-4-[193-194] bash scripts/ci_lint.sh py-lint &&
-          bash scripts/ci_lint.sh cpp-lint &&  cd ${NFS_PATH}/${GITHUB_RUN_NUMBER}/ && rm -rf ${GITHUB_JOB} || ( cd ${NFS_PATH}/${GITHUB_RUN_NUMBER}/ && rm -rf ${GITHUB_JOB} && exit 1 )
-          """
+  # Lint-C-Python:
+  #   name: Lint c python
+  #   runs-on: github-poc-ci
+  #   needs: [Rsync]
+  #   env:
+  #     MLU_REQUESTS: 0
+  #   steps:
+  #     - name: lint-c-python
+  #       run: |
+  #         ssh ${CLUSTER_CAMB} """
+  #         set -e
+  #         cd ${NFS_PATH}/ && ls -al && find ${NFS_PATH}/ -maxdepth 1 -ctime 1 -type d |xargs rm -rf
+  #         source /mnt/lustre/share/platform/env/${ENV_NAME}
+  #         cd ${NFS_PATH}/${GITHUB_RUN_NUMBER} && rm -rf ${GITHUB_JOB} && cp -R source ${GITHUB_JOB} && cd ${GITHUB_JOB}
+  #         srun --job-name=${GITHUB_JOB} --partition=${SLURM_PAR_CAMB} -x HOST-10-142-4-[193-194] bash scripts/ci_lint.sh py-lint &&
+  #         bash scripts/ci_lint.sh cpp-lint &&  cd ${NFS_PATH}/${GITHUB_RUN_NUMBER}/ && rm -rf ${GITHUB_JOB} || ( cd ${NFS_PATH}/${GITHUB_RUN_NUMBER}/ && rm -rf ${GITHUB_JOB} && exit 1 )
+  #         """
 
   Build-Pytorch:
     name: Build pytorch
@@ -112,6 +112,7 @@ jobs:
           set -e
           source /mnt/lustre/share/platform/env/${ENV_NAME}
           cd ${NFS_PATH}/${GITHUB_RUN_NUMBER}/ && rm -rf ${GITHUB_JOB} && cp -R source ${GITHUB_JOB}  && cd ${GITHUB_JOB}
+          source scripts/ci_camb_env.sh
           srun --job-name=${GITHUB_JOB} --partition=${SLURM_PAR_CAMB} -x HOST-10-142-4-[193-194] --time=40 \
           --gres=mlu:${MLU_REQUESTS} bash scripts/ci_camb_script.sh build_dipu \
           || ( cd ${NFS_PATH}/${GITHUB_RUN_NUMBER}/ && rm -rf ${GITHUB_JOB} && exit 1 )
