--- tests/test_log_softmax.py
@@ -1,18 +1,0 @@
-import torch
-import torch_dipu
-
-def test_log_softmax(x, devicestr : str):
-    device = torch.device(devicestr)
-    x = x.to(device)
-    x.requires_grad_(True)
-    log_softmax_x = torch.nn.functional.log_softmax(x, dim=1)
-    print(log_softmax_x)
-
-    grad_output = torch.ones(2, 3).to(device)
-    log_softmax_x.backward(grad_output)
-    print(x.grad)
-
-
-x = torch.randn(2, 3)
-test_log_softmax(x, "dipu")
-test_log_softmax(x, "cpu")
