+++ tests/test_ops/archived/test_adaptive_avg_pool2d.py
