+++ tests/test_ops/archived/test_batch_norm.py
