+++ tests/test_ops/archived/test_rt_tensor.py
